feat: Update sym link creation and exclusion rules
The changes in this commit refactor the symbolic link creation process in the initial zsh profile. The exclusion rule in "prezto/init.zsh" has been updated to ignore the files: "README.md", "zpreztorc" and "zshrc". Also, in "zsh/init.zsh", a symbolic link is now being created for the main zsh configuration file instead of directly sourcing it.